+++ appinfo/routes.php
@@ -97,12 +97,13 @@ $app->registerRoutes($this, ['routes' => [
 	['name' => 'podcastApi#resetAll',		'url' => '/api/podcasts/reset',					'verb' => 'POST'],
 
 	// settings API
-	['name' => 'setting#getAll',			'url' => '/api/settings',					'verb' => 'GET'],
-	['name' => 'setting#userPath',			'url' => '/api/settings/user/path',			'verb' => 'POST'],
-	['name' => 'setting#userExcludedPaths',	'url' => '/api/settings/user/exclude_paths','verb' => 'POST'],
-	['name' => 'setting#addUserKey',		'url' => '/api/settings/userkey/add',		'verb' => 'POST'],
-	['name' => 'setting#generateUserKey',	'url' => '/api/settings/userkey/generate',	'verb' => 'POST'],
-	['name' => 'setting#removeUserKey',		'url' => '/api/settings/userkey/remove',	'verb' => 'POST'],
+	['name' => 'setting#getAll',			'url' => '/api/settings',							'verb' => 'GET'],
+	['name' => 'setting#userPath',			'url' => '/api/settings/user/path',					'verb' => 'POST'],
+	['name' => 'setting#userExcludedPaths',	'url' => '/api/settings/user/exclude_paths',		'verb' => 'POST'],
+	['name' => 'setting#enableScanMetadata','url' => '/api/settings/user/enable_scan_metadata',	'verb' => 'POST'],
+	['name' => 'setting#addUserKey',		'url' => '/api/settings/userkey/add',				'verb' => 'POST'],
+	['name' => 'setting#generateUserKey',	'url' => '/api/settings/userkey/generate',			'verb' => 'POST'],
+	['name' => 'setting#removeUserKey',		'url' => '/api/settings/userkey/remove',			'verb' => 'POST'],
 
 	// Ampache API https://github.com/ampache/ampache/wiki/Ampache-API
 	['name' => 'ampache#xmlApi',	'url' => '/ampache/server/xml.server.php',	'verb' => 'GET'],